Changed Event system so that it uses event pointers instead of copy semantics.
I think this is better, because it enables us to use the Event class as a base class, and subclass it to create custom events.
Memory issues are handled by letting the EventQueue/EventLoop system "own" the event. It will be deleted after the event handler is finished with it.
This means that the "repeating timer" can not reuse the same event. Currently this feature does not work, and is therefore disabled in the test app.