Redesigned DebugStream system.
It is now much nicer code. The DebugStream class now inherits stringstream instead of using one. The template function is removed because being an ostrea, the class now have all the operators. Copy operations are totally disabled, so it is not needed to check anything in the destructor. Since it cant be copied anymore, the funcitons info, warning etc. cant return a DebugStream object anymore, and are now manipulators instead of creating the stream.

The downside of this is that a little more code is needed in each debug statement:
Before: debug() << "Hello";
After:  DebugStream() << debug << "Hello";

To avoid this, I added macros for the statement above, so that it becomes:
DEBUG << "Hello";

Yes, it is a macro, and we hate it, but I think we can live with this innocent little macro. We know that this works, and will not need to debug the debug statements.