Merged from work branch "branches/stream-targets".
A DataStream nwo always operates on an abstract "stream target".
A TCPSocket is a stream target.
The old default behaviour of the data stream is implemented in the DataBuffer stream target.
The subclass BufferStream creates an internal DataBuffer, and works just like the old DataStream.
Added File class, which is a stream target similar to TCPSocket that operates on a local file.
TCPSocket::read() never returns empty data, but throws DisconnectedException.
TCPSocket::write() is aware of disconnection, and throws DisconnectedException.
TCPSocket::write() writes to an internal buffer that must be flushed.
TCPSocket::send() writes to buffer, then flush.
TCPSocket can now be opened and reused after closed.