Designed a new draft object system.
All objects implement the abstract interface GameObject.
Users can create subclasses for their custom objects.
Users must write functions that save and restore member variables to DataStream.
Pieces maintain data as a single ByteArray for each object.

This is much better than what we have discussed so far, because users can create their own classes that takes care of both data and functionality in one place. All the user have to do in order to make processes compatible is to use the same set of classes.

What remains is to design a system for how the classes are instantiated in the peers when the host creates objects. I imagine some kind of user-customizable "object factory".