Redesigned Object class tree.
The Object class is now common for all object, whether they implement the host, peer of both interfaces.
The GameObject class is now only a convenience base class that puts together the base Object class and the host and peer interface.
Added HostGameObject and PeerGameObject that is similar to GameObject, but specialized for either process type.

Added documentation.