Updated objects test file so that it sends object data across network.
Can be started as either host (with a "host" argument) or peer. The host starts sending data right away, so the peer must be started before the host has sent 10 frames. The Peer is notifyed of incoming frame data by an event posted directly using a global pointer. This must be changed. The host is notified of the connected peer by the peer sending an event to the host, and the host exctacts the IP address, and adds the hard-coded port number 3333, to the data receivers list.